Distribute edge case tests into appropriate test files
Move tests from standalone test_coverage_improvements.py into existing
test files where they belong:

- test_flu.py: 52 edge case tests added inline to existing test functions
  - Empty iterator handling (reduce, min, max, sum, count)
  - Attribute/item access errors (map_item, map_attr)
  - Parameter edge cases (head/tail/take with 0, collect with 0)
  - __getitem__ edge cases (negative index, slices with step)
  - Side effect exception handling
  - Integration tests for complex pipelines

- test_cli.py: 4 error handling tests
  - Non-existent module import
  - Non-existent attribute import
  - Empty import list
  - Exception propagation from commands

- test_cli_utils.py: 4 utility edge case tests
  - Return type verification
  - Empty directory handling

Removed standalone coverage files in favor of inline tests.
