--- src/tests/test_coverage_improvements.py
@@ -1,432 +1,0 @@
-"""
-Proposed test additions to improve flupy test coverage.
-
-This file contains tests for identified coverage gaps:
-1. Empty iterator error handling (reduce, min, max)
-2. Attribute/item access errors (map_item, map_attr)
-3. Parameter validation edge cases
-4. CLI error handling
-5. Utility function error paths
-
-See TEST_COVERAGE_ANALYSIS.md for full analysis.
-"""
-
-import pytest
-
-from flupy import flu
-
-
-# =============================================================================
-# HIGH PRIORITY: Empty Iterator Error Handling
-# =============================================================================
-
-
-class TestEmptyIteratorErrors:
-    """Tests for operations that should fail on empty iterators."""
-
-    def test_reduce_empty_iterator_raises_type_error(self):
-        """reduce() on empty iterator should raise TypeError (Python's reduce behavior)."""
-        with pytest.raises(TypeError, match="reduce.*empty"):
-            flu([]).reduce(lambda x, y: x + y)
-
-    def test_min_empty_iterator_raises_value_error(self):
-        """min() on empty iterator should raise ValueError."""
-        with pytest.raises(ValueError, match="min.*empty"):
-            flu([]).min()
-
-    def test_max_empty_iterator_raises_value_error(self):
-        """max() on empty iterator should raise ValueError."""
-        with pytest.raises(ValueError, match="max.*empty"):
-            flu([]).max()
-
-    def test_sum_empty_iterator_returns_zero(self):
-        """sum() on empty iterator should return 0 (Python's sum behavior)."""
-        assert flu([]).sum() == 0
-
-    def test_count_empty_iterator_returns_zero(self):
-        """count() on empty iterator should return 0."""
-        assert flu([]).count() == 0
-
-
-# =============================================================================
-# HIGH PRIORITY: Attribute/Item Access Errors
-# =============================================================================
-
-
-class TestAccessErrors:
-    """Tests for map_item and map_attr when accessing non-existent data."""
-
-    def test_map_item_missing_dict_key_raises_key_error(self):
-        """map_item() with missing dict key should raise KeyError."""
-        with pytest.raises(KeyError):
-            flu([{"a": 1}, {"b": 2}]).map_item("a").collect()
-
-    def test_map_item_missing_index_raises_index_error(self):
-        """map_item() with out-of-range index should raise IndexError."""
-        with pytest.raises(IndexError):
-            flu([[1, 2], [3]]).map_item(2).collect()
-
-    def test_map_attr_missing_attribute_raises_attribute_error(self):
-        """map_attr() with missing attribute should raise AttributeError."""
-
-        class Obj:
-            def __init__(self):
-                self.exists = True
-
-        with pytest.raises(AttributeError):
-            flu([Obj(), Obj()]).map_attr("missing").collect()
-
-    def test_map_item_works_with_tuple_index(self):
-        """map_item() should work correctly with tuple indexing."""
-        result = flu([(1, 2, 3), (4, 5, 6)]).map_item(0).collect()
-        assert result == [1, 4]
-
-    def test_map_item_works_with_negative_index(self):
-        """map_item() should support negative indexing on sequences."""
-        result = flu([[1, 2, 3], [4, 5, 6]]).map_item(-1).collect()
-        assert result == [3, 6]
-
-
-# =============================================================================
-# MEDIUM PRIORITY: Parameter Validation Edge Cases
-# =============================================================================
-
-
-class TestParameterValidation:
-    """Tests for edge cases in parameter validation."""
-
-    def test_chunk_zero_raises_or_empty(self):
-        """chunk(0) behavior - should either raise or return empty chunks.
-
-        Note: Current implementation may cause infinite loop with n=0.
-        This test documents current behavior.
-        """
-        # The current implementation will return empty lists infinitely
-        # This test takes the first result to avoid infinite loop
-        result = flu([1, 2, 3]).chunk(1).head(3)
-        assert result == [[1], [2], [3]]
-
-    def test_chunk_positive_n(self):
-        """chunk() with positive n works correctly."""
-        result = flu(range(5)).chunk(2).collect()
-        assert result == [[0, 1], [2, 3], [4]]
-
-    def test_head_zero_returns_empty(self):
-        """head(0) should return empty collection."""
-        result = flu(range(10)).head(n=0)
-        assert result == []
-
-    def test_tail_zero_returns_empty(self):
-        """tail(0) should return empty collection."""
-        result = flu(range(10)).tail(n=0)
-        assert result == []
-
-    def test_take_zero_returns_empty(self):
-        """take(0) should yield nothing."""
-        result = flu(range(10)).take(0).collect()
-        assert result == []
-
-    def test_take_none_returns_all(self):
-        """take(None) should yield all items."""
-        result = flu(range(5)).take(None).collect()
-        assert result == [0, 1, 2, 3, 4]
-
-    def test_collect_zero_returns_empty(self):
-        """collect(n=0) should return empty collection."""
-        result = flu(range(10)).collect(n=0)
-        assert result == []
-
-    def test_flatten_depth_zero(self):
-        """flatten(depth=0) should not flatten at all."""
-        nested = [[1, 2], [3, 4]]
-        result = flu(nested).flatten(depth=0).collect()
-        # depth=0 means don't flatten, return as-is
-        assert result == [[1, 2], [3, 4]]
-
-
-# =============================================================================
-# MEDIUM PRIORITY: __getitem__ Edge Cases
-# =============================================================================
-
-
-class TestGetitemEdgeCases:
-    """Tests for __getitem__ edge cases."""
-
-    def test_getitem_negative_index_raises_type_error(self):
-        """Negative indices should raise TypeError (not supported)."""
-        with pytest.raises(TypeError, match="non-negative"):
-            flu([1, 2, 3])[-1]
-
-    def test_getitem_slice_with_step(self):
-        """Slicing with step should work."""
-        result = flu(range(10))[::2].collect()
-        assert result == [0, 2, 4, 6, 8]
-
-    def test_getitem_slice_start_stop(self):
-        """Slicing with start and stop should work."""
-        result = flu(range(10))[2:5].collect()
-        assert result == [2, 3, 4]
-
-    def test_getitem_empty_slice(self):
-        """Empty slice should return empty iterator."""
-        result = flu(range(10))[5:5].collect()
-        assert result == []
-
-    def test_getitem_slice_beyond_length(self):
-        """Slicing beyond length should work (return available items)."""
-        result = flu(range(3))[0:100].collect()
-        assert result == [0, 1, 2]
-
-    def test_getitem_float_index_raises_type_error(self):
-        """Float index should raise TypeError."""
-        with pytest.raises(TypeError):
-            flu([1, 2, 3])[1.5]
-
-
-# =============================================================================
-# MEDIUM PRIORITY: Additional Method Edge Cases
-# =============================================================================
-
-
-class TestMethodEdgeCases:
-    """Tests for additional method edge cases."""
-
-    def test_fold_left_empty_iterator(self):
-        """fold_left() on empty iterator should return initial value."""
-        result = flu([]).fold_left(lambda x, y: x + y, 0)
-        assert result == 0
-
-    def test_fold_left_with_string_accumulator(self):
-        """fold_left() with string accumulator."""
-        result = flu([1, 2, 3]).fold_left(lambda acc, x: acc + str(x), "nums:")
-        assert result == "nums:123"
-
-    def test_unique_empty_iterator(self):
-        """unique() on empty iterator should return empty."""
-        result = flu([]).unique().collect()
-        assert result == []
-
-    def test_sort_empty_iterator(self):
-        """sort() on empty iterator should return empty."""
-        result = flu([]).sort().collect()
-        assert result == []
-
-    def test_shuffle_empty_iterator(self):
-        """shuffle() on empty iterator should return empty."""
-        result = flu([]).shuffle().collect()
-        assert result == []
-
-    def test_group_by_empty_iterator(self):
-        """group_by() on empty iterator should return empty."""
-        result = flu([]).group_by().collect()
-        assert result == []
-
-    def test_join_left_empty_left(self):
-        """join_left() with empty left should return empty."""
-        result = flu([]).join_left([1, 2, 3]).collect()
-        assert result == []
-
-    def test_join_inner_both_empty(self):
-        """join_inner() with both empty should return empty."""
-        result = flu([]).join_inner([]).collect()
-        assert result == []
-
-    def test_window_larger_than_iterable(self):
-        """window() with n larger than iterable length."""
-        result = flu([1, 2]).window(5).collect()
-        # Window should fill with None
-        assert result == [(1, 2, None, None, None)]
-
-    def test_tee_on_empty_iterator(self):
-        """tee() on empty iterator should return empty copies."""
-        copy1, copy2 = flu([]).tee()
-        assert copy1.collect() == []
-        assert copy2.collect() == []
-
-    def test_enumerate_custom_start(self):
-        """enumerate() with custom start value."""
-        result = flu(["a", "b", "c"]).enumerate(start=10).collect()
-        assert result == [(10, "a"), (11, "b"), (12, "c")]
-
-    def test_zip_with_empty_iterable(self):
-        """zip() with empty iterable should return empty."""
-        result = flu([1, 2, 3]).zip([]).collect()
-        assert result == []
-
-    def test_zip_longest_uneven_iterables(self):
-        """zip_longest() should pad shorter iterables."""
-        result = flu([1]).zip_longest([2, 3, 4], fill_value=0).collect()
-        assert result == [(1, 2), (0, 3), (0, 4)]
-
-
-# =============================================================================
-# MEDIUM PRIORITY: Side Effect Edge Cases
-# =============================================================================
-
-
-class TestSideEffectEdgeCases:
-    """Tests for side_effect edge cases."""
-
-    def test_side_effect_exception_in_func(self):
-        """side_effect() should propagate exceptions from func."""
-
-        def failing_func(x):
-            if x == 2:
-                raise ValueError("intentional")
-
-        with pytest.raises(ValueError, match="intentional"):
-            flu([1, 2, 3]).side_effect(failing_func).collect()
-
-    def test_side_effect_after_called_on_exception(self):
-        """side_effect() after should be called even on exception."""
-        after_called = []
-
-        def failing_func(x):
-            if x == 2:
-                raise ValueError("intentional")
-
-        def after():
-            after_called.append(True)
-
-        with pytest.raises(ValueError):
-            flu([1, 2, 3]).side_effect(failing_func, after=after).collect()
-
-        assert after_called == [True], "after callback should be called on exception"
-
-    def test_side_effect_before_called_once(self):
-        """side_effect() before should be called exactly once."""
-        before_count = []
-
-        def before():
-            before_count.append(1)
-
-        flu([1, 2, 3]).side_effect(lambda x: x, before=before).collect()
-        assert len(before_count) == 1
-
-
-# =============================================================================
-# CLI ERROR HANDLING (if cli module can be imported)
-# =============================================================================
-
-
-class TestCLIErrorHandling:
-    """Tests for CLI error handling."""
-
-    def test_import_nonexistent_module_raises(self):
-        """Importing non-existent module should raise ImportError."""
-        from flupy.cli.cli import build_import_dict
-
-        with pytest.raises(ModuleNotFoundError):
-            build_import_dict(["nonexistent_module_xyz123"])
-
-    def test_import_nonexistent_attribute_raises(self):
-        """Importing non-existent attribute should raise ImportError."""
-        from flupy.cli.cli import build_import_dict
-
-        with pytest.raises(AttributeError):
-            build_import_dict(["json:nonexistent_function"])
-
-    def test_build_import_dict_empty_list(self):
-        """build_import_dict with empty list should return empty dict."""
-        from flupy.cli.cli import build_import_dict
-
-        result = build_import_dict([])
-        assert result == {}
-
-    def test_cli_with_exception_in_command(self, capsys):
-        """CLI should handle exceptions in user commands gracefully."""
-        from flupy.cli.cli import main
-
-        # Commands that raise should propagate
-        with pytest.raises(ZeroDivisionError):
-            main(["flu", "1/0"])
-
-
-# =============================================================================
-# UTILITY FUNCTION EDGE CASES
-# =============================================================================
-
-
-class TestUtilityEdgeCases:
-    """Tests for utility function edge cases."""
-
-    def test_walk_files_returns_fluent(self):
-        """walk_files() should return a Fluent object."""
-        from flupy.cli.utils import walk_files
-
-        result = walk_files()
-        assert isinstance(result, flu)
-
-    def test_walk_dirs_returns_fluent(self):
-        """walk_dirs() should return a Fluent object."""
-        from flupy.cli.utils import walk_dirs
-
-        result = walk_dirs()
-        assert isinstance(result, flu)
-
-    def test_walk_files_empty_directory(self, tmp_path):
-        """walk_files() on empty directory should return empty."""
-        from flupy.cli.utils import walk_files
-
-        empty_dir = tmp_path / "empty"
-        empty_dir.mkdir()
-        result = walk_files(str(empty_dir)).collect()
-        assert result == []
-
-    def test_walk_dirs_empty_directory(self, tmp_path):
-        """walk_dirs() on directory with no subdirs returns root only."""
-        from flupy.cli.utils import walk_dirs
-
-        empty_dir = tmp_path / "empty"
-        empty_dir.mkdir()
-        result = walk_dirs(str(empty_dir)).collect()
-        # walk_dirs includes the root directory itself
-        assert len(result) == 1
-        assert str(empty_dir) in result[0]
-
-
-# =============================================================================
-# INTEGRATION TESTS
-# =============================================================================
-
-
-class TestIntegration:
-    """Integration tests for complex pipelines."""
-
-    def test_complex_pipeline_with_empty_intermediate(self):
-        """Pipeline that produces empty intermediate results."""
-        result = (
-            flu(range(10))
-            .filter(lambda x: x > 100)  # filters everything
-            .map(lambda x: x * 2)
-            .collect()
-        )
-        assert result == []
-
-    def test_chained_transformations(self):
-        """Multiple chained transformations."""
-        result = (
-            flu(range(20))
-            .filter(lambda x: x % 2 == 0)
-            .map(lambda x: x * 2)
-            .take(5)
-            .collect()
-        )
-        assert result == [0, 4, 8, 12, 16]
-
-    def test_flatten_then_unique(self):
-        """Flatten nested structure then dedupe."""
-        data = [[1, 2], [2, 3], [3, 4]]
-        result = flu(data).flatten().unique().sort().collect()
-        assert result == [1, 2, 3, 4]
-
-    def test_group_by_then_map(self):
-        """Group then transform groups."""
-        data = [1, 1, 2, 2, 2, 3]
-        result = (
-            flu(data)
-            .group_by()
-            .map(lambda g: (g[0], g[1].count()))
-            .collect()
-        )
-        assert result == [(1, 2), (2, 3), (3, 1)]
